[#1] Refactor secret input handling: replace SecretInput and SecretOptions with SecretFormLite component
